+++ aerofiles/seeyou/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 
-from .common import WaypointStyle, ObservationZoneStyle
+from .common import WaypointStyle, ObservationZoneStyle, SeeYouFileFormat
 from .converter import Converter
 from .reader import Reader
 from .writer import Writer